misc/HtmlTestSuite/results/test-0149.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf8"/>
<title>CSS contain-intrinsic-size: scroll containers</title>
<link href="mailto:obrufau@igalia.com" rel="author" title="Oriol Brufau"/>
<link href="https://drafts.csswg.org/css-sizing-4/#intrinsic-size-override" rel="help"/>
<link href="https://drafts.csswg.org/css-overflow-3/#scroll-container" rel="help"/>
<meta content="width=device-width, initial-scale=1" name="viewport"/>
<style>
.test {
  contain: size;
  display: inline-block;
  padding: 0;
  border: 5px solid;
}
.test::before {
  content: '';
  display: block;
  width: 40px;
  height: 20px;
}
.big-contents::before {
  width: 400px;
  height: 200px;
}
.overflow-auto {
  overflow: auto;
  scrollbar-gutter: stable;
}
.overflow-scroll {
  overflow: scroll;
}
.overflow-hidden {
  overflow: hidden;
}
.cis-none {
  contain-intrinsic-size: none none;
}
.cis-height {
  contain-intrinsic-size: none 50px;
}
.cis-width {
  contain-intrinsic-size: 100px none;
}
.cis-both {
  contain-intrinsic-size: 100px 50px;
}
</style>

</head><body><div id="log"/>

<div class="test small-contents overflow-auto cis-none" data-expected-client-height="0" data-expected-client-width="0" data-expected-scroll-height="20" data-expected-scroll-width="40"/>
<div class="test small-contents overflow-auto cis-height" data-expected-client-height="50" data-expected-client-width="0" data-expected-scroll-height="50" data-expected-scroll-width="40"/>
<div class="test small-contents overflow-auto cis-width" data-expected-client-height="0" data-expected-client-width="100" data-expected-scroll-height="20" data-expected-scroll-width="100"/>
<div class="test small-contents overflow-auto cis-both" data-expected-client-height="50" data-expected-client-width="100" data-expected-scroll-height="50" data-expected-scroll-width="100"/>

<div class="test small-contents overflow-scroll cis-none" data-expected-client-height="0" data-expected-client-width="0" data-expected-scroll-height="20" data-expected-scroll-width="40"/>
<div class="test small-contents overflow-scroll cis-height" data-expected-client-height="50" data-expected-client-width="0" data-expected-scroll-height="50" data-expected-scroll-width="40"/>
<div class="test small-contents overflow-scroll cis-width" data-expected-client-height="0" data-expected-client-width="100" data-expected-scroll-height="20" data-expected-scroll-width="100"/>
<div class="test small-contents overflow-scroll cis-both" data-expected-client-height="50" data-expected-client-width="100" data-expected-scroll-height="50" data-expected-scroll-width="100"/>

<div class="test small-contents overflow-hidden cis-none" data-expected-client-height="0" data-expected-client-width="0" data-expected-scroll-height="20" data-expected-scroll-width="40"/>
<div class="test small-contents overflow-hidden cis-height" data-expected-client-height="50" data-expected-client-width="0" data-expected-scroll-height="50" data-expected-scroll-width="40"/>
<div class="test small-contents overflow-hidden cis-width" data-expected-client-height="0" data-expected-client-width="100" data-expected-scroll-height="20" data-expected-scroll-width="100"/>
<div class="test small-contents overflow-hidden cis-both" data-expected-client-height="50" data-expected-client-width="100" data-expected-scroll-height="50" data-expected-scroll-width="100"/>


<div class="test big-contents overflow-auto cis-none" data-expected-client-height="0" data-expected-client-width="0" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-auto cis-height" data-expected-client-height="50" data-expected-client-width="0" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-auto cis-width" data-expected-client-height="0" data-expected-client-width="100" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-auto cis-both" data-expected-client-height="50" data-expected-client-width="100" data-expected-scroll-height="200" data-expected-scroll-width="400"/>

<div class="test big-contents overflow-scroll cis-none" data-expected-client-height="0" data-expected-client-width="0" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-scroll cis-height" data-expected-client-height="50" data-expected-client-width="0" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-scroll cis-width" data-expected-client-height="0" data-expected-client-width="100" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-scroll cis-both" data-expected-client-height="50" data-expected-client-width="100" data-expected-scroll-height="200" data-expected-scroll-width="400"/>

<div class="test big-contents overflow-hidden cis-none" data-expected-client-height="0" data-expected-client-width="0" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-hidden cis-height" data-expected-client-height="50" data-expected-client-width="0" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-hidden cis-width" data-expected-client-height="0" data-expected-client-width="100" data-expected-scroll-height="200" data-expected-scroll-width="400"/>
<div class="test big-contents overflow-hidden cis-both" data-expected-client-height="50" data-expected-client-width="100" data-expected-scroll-height="200" data-expected-scroll-width="400"/>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<script>
checkLayout(".test");
</script>
</body></html>
